+++ eox_audit_model/tests/tests_models.py
@@ -69,9 +69,9 @@ class TestGetCurrentSite(TestCase):
 class TestGetCurrentIp(TestCase):
     """Test cases for get_current_site."""
 
-    @patch('eox_audit_model.models.get_ip')
+    @patch('eox_audit_model.models.ip')
     @patch('eox_audit_model.models.get_current_request')
-    def test_valid_request(self, get_current_request_mock, get_ip_mock):
+    def test_valid_request(self, get_current_request_mock, ip_mock):
         """This method tests when the request is not None.
 
         Expected behavior:
@@ -81,18 +81,40 @@ class TestGetCurrentIp(TestCase):
         """
         request = Mock()
         expected_value = '192.163.45.67'
-        get_ip_mock.return_value = expected_value
+        ip_mock.get_ip.return_value = expected_value
         get_current_request_mock.return_value = request
 
         result = get_current_ip()
 
         self.assertEqual(expected_value, result)
         get_current_request_mock.assert_called_once()
-        get_ip_mock.assert_called_once_with(request)
+        ip_mock.get_ip.assert_called_once_with(request)
 
-    @patch('eox_audit_model.models.get_ip')
+    @patch('eox_audit_model.models.ip')
     @patch('eox_audit_model.models.get_current_request')
-    def test_invalid_request(self, get_current_request_mock, get_ip_mock):
+    def test_valid_request_with_get_client(self, get_current_request_mock, ip_mock):
+        """This method tests get_client_ip for django-ipware major or equals to 3.0.0
+
+        Expected behavior:
+            - Return expected value.
+            - get_current_request is called once.
+            - get_client_ip is called once.
+        """
+        request = Mock()
+        expected_value = '192.163.45.67'
+        ip_mock.get_ip.side_effect = AttributeError()
+        ip_mock.get_client_ip.return_value = (expected_value, )
+        get_current_request_mock.return_value = request
+
+        result = get_current_ip()
+
+        self.assertEqual(expected_value, result)
+        get_current_request_mock.assert_called_once()
+        ip_mock.get_client_ip.assert_called_once_with(request)
+
+    @patch('eox_audit_model.models.ip')
+    @patch('eox_audit_model.models.get_current_request')
+    def test_invalid_request(self, get_current_request_mock, ip_mock):
         """This method tests when the request is None.
 
         Expected behavior:
@@ -107,7 +129,7 @@ class TestGetCurrentIp(TestCase):
 
         self.assertEqual(expected_value, result)
         get_current_request_mock.assert_called_once()
-        get_ip_mock.assert_not_called()
+        ip_mock.get_ip.assert_not_called()
 
 
 class TestGetCurrentPerformer(TestCase):